Handle expired tokens from the ProtectedRoute component. (#104)
Previously, ProtectedRoute would load the page even if the local storage
contained an expired token. When the nested component tried to make a
request to the API, the `getBearerToken` function would check the expiry
time and would cause a redirection to the login page by assigning to
`window.location.href`.

Assigning to `window.location.href` directly does not play well with the
rest of the React router. In particular it does not take the PUBLIC_URL
environment variable into account. This means that on an expired token,
the user was always redirected to `/login` even when it should have been
`$PUBLIC_URL/login`.

Using the `navigate(...)` function works much better. However it cannot
be called directly from `getBearerToken()`, which is not a React
component. Instead it can be done from the `ProtectedRoute`. In fact
this is already implemented for the case where the user has no token at
all. All that was required was a stricter definition of the
`isAuthenticated()` function.

This change also makes sure the URL the user was trying to load with the
expired token is saved through the existing `RedirectOnLoginProvider`.
This was correctly implemented for GitHub authentication but not for the
`BasicUserAuthForm`.

Steps to reproduce:
1. Start the frontend at a subdirectory, with `PUBLIC_URL=/app npm start`
2. Log in to the application with the super-user account.
3. Manually cause the token in the local storage to be expired, via the
   DevTools console:

    ```js
    localStorage.setItem("user", JSON.stringify({
      ...JSON.parse(localStorage.getItem("user")),
      exp: 0,
    }));
    ```

4. Navigate to a protected URL, such as `/app/manage-roles`.

Step 4 should direct the user to `/app/login`. After a successful login,
the app should redirect to the `/app/manage-roles` page.

Without this change, the user is redirected to `/login`, which is not a
valid route.

In the long term I think the `getBearerToken()` function is a bit of a
code smell. It works completely outside of the rest of the application
structure, and there is no reason to load from the local storage
multiple times. Instead I think we should retrieve the token from the
`UserProvider`. This does however entail a much larger refactor than
this simple fix.