Add an option to specify default roles for a user.
The list of roles will be added to any newly created users. This
includes basic auth users created via the admin portal and implicitly
created GitHub users. The list of roles is specified in the application
properties. By default no roles are added.

Default roles aren't retroactive. If a user already exists and a default
role is configured, the users don't get the new default role. Similarly,
one can remove a default role that was granted to a user and they won't
automatically have it back.

Unknown roles, which are listed by name in the configuration but don't
exist in the database, are ignored. The goal is to simplify setting up a
new server, where one would start the server with a list of default
roles pre-configured but an empty database. The initial users then logs
in, is manually granted permission by editing the database, and may then
create the roles through the admin panel. If unknown default roles led
to a error on account creation it would be impossible for the initial
user to log in. One would have to start the server without default
roles, do the initial setup in the panel, then stop and restart the
server with the new options.