Catch exceptions when checking for org and team membership.
When we check for org and team membership, we may get a 403 error if the
GitHub token does not have sufficient permissions. This can happen quite
easily if a user creates a PAT without the right scopes. Without this,
Packit was returning empty error responses, with no body attached to it.