Add an Actions workflow to check that migrations aren't being changed.
Changing existing migrations causes all sorts of problems when deploying
to an existing server. Flyway will complain that the migration's
checksum has changed, and will require manual intervention to fix it,
either by dropping the whole database, or by manually modifying the
schema to match the new version of the migration. This defeats the
purpose of migrations, which should prevent manual management of the
schema.

This has already happened multiple times, so prevent any further
incidents by adding a workflow on PRs to prevent it. The workflow is
intentionally separate from the others to make it easier to ignore in
case of false positives.