Expose metrics over HTTP.
This enables Spring Boot Actuator, which exposes lots of monitoring
information about the service. In particular, this exposes metrics in a
Prometheus-compatible format. There are a lot of built-in metrics that
come with Spring already. In addition, this add a metric with the Git
revision of the repo that was used at build time, similar to what is
done in outpack_server.

To avoid exposing everything over the default HTTP port, the management
interface is bound to a separate port, 8081 by default.