Add a `/user` endpoint to the API.
Previously the frontend got the list of users by listing the roles and
filtering on the isUsername flag. This was a little awkward, and meant
that extra information about the user isn't accessible to the frontend.

As part of mrc-5763, I want to add a "service account", which will need
to be displayed differently, or maybe won't even be displayed at all.
Because the endpoint the front end is calling doesn't get the full user
details, it cannot tell whether a user is a service account or not.

For this to work, a new `GET /user` endpoint is added, which lists the
users and all their properties. The frontend is modified to use this new
endpoint. Both the `/role` and the `/user` endpoints are now setup to
never return username roles. From the perspective of the frontend, these
special roles don't exist anymore. The `/user` endpoint includes the
permissions that are specific to the user.

There's a lot of churn in the frontend from renaming some files and
classes into something I feel makes more sense.

I've removed a few bits of code that were dead, especially in the User
and Role repositories.