Add secret-less authentication for CI jobs.
This change allows automated environments, such as Github Actions or
Buildkite to authenticate to the Packit API without needing to be
manually provisioned with a secret.

The client's environment creates and signs a JWT containing claims
describing the job. For example, on Github Actions, GitHub provides a
token which includes the organisation name, repository name, type of
event, branch name, etc. Packit needs to be configured to trust the 3rd
party JWT issuer, with policies specifying what claims are needed.

The client sends the JWT it received from its environment to the Packit
`/auth/login/jwt` endpoint, and in exchange receives a Packit JWT which
may be used to access the rest of the API.