Update to new orderly.runner interface.
The new runner does not share an on-disk repository with outpack server
anymore. Instead it uses it supports an arbitrary number of
repositories, and each request needs to provide the repository URL.

For now, a most one Git repository can configured in the application
properties. In the future, we may expand this to multiple repositories,
or even allow them to be configured from the frontend and stored in the
database.

When submitting a run request, we also need to provide details of the
outpack location. We use the same URL to outpack_server that Packit uses
itself. For now at least, Packit, outpack_server and the runner workers
all run on the same machine. In the future, we may make it so that the
runner workers communicate via Packit, and add an authentication
mechanism.