Tidy up role controller endpoints.
This renames the role endpoint from `/role` to `/roles`, bringing it
inline with the existing `/packets` and `/packetGroups`.

This also removes the `GET /roles/names` endpoint, which was not being
used anywhere, and which overlapped with the existing
`GET /roles/<name>` which is used to get details about a particular
role. Having both routes could cause issue is a role was named `names`.