Tidy up packet and packet group controllers.
The main motivation for this change was to remove some ambiguity in the
old routes. In particular, we had both the `/packets/<name>` and
`/packets/packetGroupSummaries` routes.

Some of the PacketController routes had more to do with packet groups
and have been moved to the appropriate controller.

The PacketController routes had `/metadata/<id>` and `/file/<id>`
routes, which have now been replaced by `/<id>` and `/<id>/file`, which
I think is more consistent with the rest of the API.

This is an overview of the routes after these changes:
- `/packets`: returns a list of all packets, searchable by id / name.
- `/packets/<id>`: returns a packet's metadata
- `/packets/<id>/file`: returns a file from a packet
- `/packetGroups`: returns the list of packet groups, searchable by
  name. Only returns the id and name of the group.
- `/packetGroups/<name>/display`: returns high-level information of the group,
  ie. the display name and description.
- `/packetGroups/<name>/packets`: returns the list of packets in that
  group.
- `/packetGroupSummaries`: returns a list of packet groups, searchable
  by name and display name. Includes a "summary" for each group,
  including the number of packets and the latest packet ID.

I find the duplication between `/packetGroups` and
`/packetGroupSummaries` a bit unfortunate. I think these should be
merged into a single endpoint. I have left this for a future PR.

The `/packets` and `/packetGroups` endpoints are a bit special in that
they are used by the frontend way in a generic way to set scoped
permissions, therefore must have a relatively uniform interface
(together with the `/tag` endpoint).