Implement the sd_notify protocol.
systemd is the standard service manager used on Linux systems, and the
one we use on our production machines.

It is able to track the state of services, including whether a service
is "starting" or "started" already. For most services that are quick to
start, the difference is minimal but in the case of Packit it can take a
handful of seconds or more for the service to be ready to accept
connections.

Thankfully systemd provides an easy way to notify it of the service's
state. If the service is configured with `Type=notify`, systemd will set
a `NOTIFY_SOCKET` environment variable point to a Unix socket. The
service can communicate with systemd over that socket, and in particular
it can write `READY=1` to change the service's state from "starting" to
"started". See [the systemd documentation][sd_notify] for more details
about the protocol and the various messages it supports.

Having systemd know accurately when the service has started allows us to
coordinate startup more easily (for example we could execute some
provisioning steps automatically once the service has started). It can
also make whole-system integration tests easier to write, by providing
an easy way to start the service and wait for it to be ready, without
having to repeatedly poll the HTTP endpoint.

Testing this change outside of the context of a systemd service is a
little tricky. The best way to reproduce this is to run packit as a
transient unit with `systemd-run`:

```sh
# Build the app's JAR file
gradle :app:bootJar

# Start Packit as a transient unit, telling systemd we support the
# notify protocol
systemd-run --user --collect --service-type=notify --same-dir \
  --unit=packit -- java -jar app/build/libs/app.jar

# Watch the logs for updates from systemd about the state of the
# service.
journalctl --user --unit packit.service

# Cleanup
systemctl --user stop packit.service
```

The logs show systemd updating the status of the `packit.service` unit
just after Spring has completed initialization.

```
14:35:08 systemd[2970]: Starting packit.service
14:35:09 java[276126]:   .   ____          _            __ _ _
14:35:09 java[276126]:  /\\ / ___'_ __ _ _(_)_ __  __ _ \ \ \ \
14:35:09 java[276126]: ( ( )\___ | '_ | '_| | '_ \/ _` | \ \ \ \
14:35:09 java[276126]:  \\/  ___)| |_)| | | | | || (_| |  ) ) ) )
14:35:09 java[276126]:   '  |____| .__|_| |_|_| |_\__, | / / / /
14:35:09 java[276126]:  =========|_|==============|___/=/_/_/_/
14:35:09 java[276126]:  :: Spring Boot ::                (v3.0.2)
[...] Lots more output from spring
14:35:14 java[276126]: Tomcat started on port(s): 8080 (http) with context path ''
14:35:14 java[276126]: Started AppKt in 5.372 seconds (process running for 5.771)
14:35:14 systemd[2970]: Started packit.service
```

Amazingly, systemd has actually added an option just yesterday that
makes it easier to start programs that support sd_notify, without
needing to create a transient service:
https://github.com/systemd/systemd/commit/4389e4c2aea36224e99dae616aff08c8560c94e8
Unfortunately, it will be a little while until we can actually make use
of this.

[sd_notify]: https://www.freedesktop.org/software/systemd/man/latest/sd_notify.html