Add a 'wait-notify' script.
This implements the supervisor-side of the sd_notify protocol.

It can be used either with local processes, or as a wrapper around
`docker run`:

```sh
./scripts/wait-notify -- java -jar api/app/build/libs/app.jar
./scripts/wait-notify --docker-run -- mrcide/packit-api
```

This has been integrated in the scripts responsible for starting the API
container, removing the need to poll and wait for the API to become
ready.

Ideally we would also use this to start the database container, and
remove the `wait-for-db` script. Unfortunately the Postgres image has
some quirk that makes it incompatible, despite Postgres natively
supporting sd_notify. The first time the container starts, it runs a
temporary server used to initialiase the database storage. This
temporary server is not accesible, but because it inherits the
NOTIFY_SOCKET environment variable, it notifies the caller and the
container appears ready too early. This can be fixed upstream in the
postgres image.