Use TIMESTAMPTZ column instead of TIMESTAMP column so that times normalize to UTC
When we passed a timestamp in UTC (from Instant.now) to be stored in a database column, it was being converted into BST at the point of storing, which is one hour off from UTC. This was ultimately a result of using the wrong PostgreSQL type for the database column.

PostgreSQL provides two main timestamp data types, TIMESTAMP and TIMESTAMPTZ (short for ‘timestamp with time zone’ - where the time zone is always UTC.)

The main difference in definition is that TIMESTAMPTZ is storing the timestamp in UTC, while TIMESTAMP makes no claims about what timezone it’s in.

The main differences in behaviour are (1) TIMESTAMPTZ columns convert values into UTC before storing them, and (2) TIMESTAMPTZ values are automatically changed if the database timezone is changed, while the TIMESTAMP ones do not.

Here are some reasons to always prefer TIMESTAMPTZ over TIMESTAMP: https://justatheory.com/2012/04/postgres-use-timestamptz/ (The section about partitioning is out of date.)

Here is how Spring Data JPA maps Java classes to PostgreSQL types: https://www.baeldung.com/java-postgresql-store-date-time\#1-default-mapping

‘Instant’ is mapped to ‘TIMESTAMP WITH TIME ZONE’, meaning that JPA probably expects the database to be using that type. That could explain why, when we asked it to save 3pm (in UTC) to a column with type ‘TIMESTAMP’ (without time zone), it stored it as 4pm instead: some middleware along the way (inside or outside Postgres, I don’t know) may have noticed that its own time zone was BST, and that the database column type is TIMESTAMP without timezone, and ‘helpfully’ converted the UTC value into a BST version before it was stored.