Move one-time token validation logic into the security module
Instead of handling validation in packet controller and OTT service, use Spring Security to
store the token's details against the security context (and delete token from database), and
use PreAuthorize annotations to validate the token.

I also changed the point in the filter chain at which the existing JWTAuthenticationFilter
is inserted, from just before the UsernamePasswordAuthenticationFilter, to just after the
LogoutFilter. This is because I wanted to follow the heuristics listed here:
https://docs.spring.io/spring-security/reference/servlet/architecture.html#_adding_a_custom_filter

This commit also changes the http status codes returned on requesting real files without
relevant permissions, from 404 to 401. 404 had been an overzealous attempt at a higher level
of security than we actually need; returning 401 for all invalid tokens,
including those where the permitted file paths don't match the requested ones, means that we
can put all authorization logic into a single function in AuthorizationLogic.kt. (We do not
directly control the status codes that result from AuthorizationLogic.kt deeming a permission
to be missing; I think it may be possible to affect this implicitly, but that is out of scope
for this work.)