Redesign header/breadcrumb layout
The motivation for these changes is that configurable logos may well contain text,
and in fact VIMC's does, and this text is unlikely to happen to align with text in
the header.

These changes seek to make it such that the user's eye does not expect to see any
alignment, by changing the hierarchy of components: now, the breadcrumb is
contained inside the header instead of outside.

At the same time, it seemed prudent to avoid the mis-alignment of the app name text
with the nav menu link to 'Runner'; this was achieved by consolidating the two nav
menus within our header into a single nav menu, which decides whether to be central
inside the header or right-aligned depending on its number of items.

The breadcrumb component now starts with an extra chevron so that when you are on
the home page, the word 'home' displayed on its own by the breadcrumb component is
easier to interpret and looks more deliberate.

In order to allow the developer to configure which pages show breadcrumbs and which
do not by use of the Router, I add a 'BreadcrumbRoute' component, which uses a
React provider to (1) set the showBreadcrumbs property to true when it mounts, and
(2) set it false when it unmounts. Setting it false when it unmounts allows us to
implicitly configure non-breadcrumb-showing pages without having to wrap those in
something too, since navigating away from a page wrapped by a BreadcrumbRoute will
consequently un-show the breadcrumbs.

Some example logos are added for developers' convenience; these can be applied by
setting packit.branding.logoFilename in application.properties.