Provide base for absolute URLs for download URL utils
These functions that calculate URLs for requesting download-related entities
were broken when deployed, because they were assuming that the configured
'api url' is absolute (which it is in development/testing).

The interface of URL requires absolute URLs, so we need to pass in a base for
it to relativise the path against.