Validate packet group name in front-end urls
At the moment you can go to https://reside.packit.dide.ic.ac.uk/this-is-not-a-packet-group-name/20250214-210242-0856dd0b and the app does not complain about the incorrect packet group name in the url - it just accepts it at face value and presents the packet as being a member of this non-existent group.

Instead we should match the name in the packet metadata with the name in the url, and display a generic looking error if they don't match.

In this commit I also refactor PacketOutlet to use the url params rather than be passed props.
This removes the need for the wrapper layout PacketBlankLayout as we can just use PacketOulet
directly now.