Refactor fetching dependency packets' info
Use an endpoint that takes a list of ids (for which we want to get the basic details)
rather than a dependencies-specific endpoint (which would use an extra metadata request).
Since we pass a list of ids that can be arbitrarily long, we use the request body
to pass it to the endpoint. This requires not using a GET request, so we use POST.